move tree actions menu component
