+++ src/frontend/src/views/Data/components/TreesView/components/TreesTable/components/TreeActionMenu/components/OpenInNextstrainButton/index.tsx
@@ -1,7 +1,7 @@
 import { ButtonIcon, Icon, Tooltip } from "czifui";
 import { useState } from "react";
 import { TREE_STATUS } from "src/common/constants/types";
-import NextstrainConfirmationModal from "../../../NextstrainConfirmationModal";
+import NextstrainConfirmationModal from "../../../../../../../NextstrainConfirmationModal";
 
 interface Props {
   item: PhyloRun;
